requirements.txt
boto3>=1.28.0
botocore>=1.31.0
h5py>=3.9.0
openslide-python>=1.3.1
Pillow>=9.0.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
